requirements.txt
asgiref==3.2.7
Django
django-widget-tweaks==1.4.8
djongo==1.3.7
gunicorn==23.0.0
packaging
pillow
pymongo==3.11.4
PyMySQL==1.1.1
python-dotenv
pytz==2020.1
sqlparse==0.2.4
whitenoise
django-import-export # Added this line
django-admin-interface